requirements.txt
beautifulsoup4
imagehash
lxml
Pillow
requests
selenium
